requirements.txt
Flask
Flask-SQLAlchemy
mysqlclient
requests
beautifulsoup4
Flask-Login
openpyxl
pillow
numpy
python-docx
docxtpl
docx2pdf